refactor utility fns into new module, create fixture for loading test data files
